server/requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-filter==1.1.0
djangorestframework==3.8.2
gunicorn==19.8.0
pytz==2018.4
